Added looping video

+++ master/video/omx-test4.py
@@ -8,19 +8,27 @@ from video_db import *
 
 DEBUG = True
 
-inter_video_delay = 0.5
+inter_video_delay = 0.75
 
-CONTENT_CMD = ['omxplayer', '--no-osd', '--no-keys', '--refresh', 
-        '--aspect-mode fill', '--layer 2']
-TRANSITION_CMD = ['omxplayer', '--no-osd', '--no-keys', '--refresh', 
-        '--aspect-mode fill', '--layer 5']
-LOOP_CMD = ['omxplayer', '--no-osd', '--no-keys', '--refresh', '--loop', '--aspect-mode fill']
+#OMX_CMD = ['omxplayer', '--no-osd', '--no-keys', '--refresh', '--aspect-mode fill']
+OMX_CMD = ['omxplayer', '--no-osd', '--no-keys', '--aspect-mode fill']
+#CONTENT_CMD = OMX_CMD + ['--layer 4']
+#TRANSITION_CMD = OMX_CMD + ['--layer 5']
+#LOOP_CMD = OMX_CMD + ['--layer 1', '--loop']
+CONTENT_CMD = OMX_CMD + ['--layer 4', '--dbus_name', 'org.mpris.MediaPlayer2.omxplayer1']
+TRANSITION_CMD = OMX_CMD + ['--layer 5', '--dbus_name', 'org.mpris.MediaPlayer2.omxplayer2']
+LOOP_CMD = OMX_CMD + ['--layer 1', '--loop', '--dbus_name', 'org.mpris.MediaPlayer2.omxplayer3']
 
 transition_film_list = []
 content_film_list = []
+loop_film_list = []
 
+pgid_last_loop = None
 pgid_last_content = None
 pgid_last_transition = None
+
+nullin = open(os.devnull, 'r')
+nullout = open(os.devnull, 'w')
 
 last_caller = ""
 def debug(*args):
@@ -42,10 +50,8 @@ def create_film_lists():
             transition_film_list.append(film)
         elif film['type'] == 'content':
             content_film_list.append(film)
-
-
-def start_default_film(film):
-    pass
+        elif film['type'] == 'loop':
+            loop_film_list.append(film)
 
 
 def start_transition(film):
@@ -55,18 +61,14 @@ def start_transition(film):
     debug("start:", film['start'], "sec")
     debug("end:", film['start']+film['length'], "sec")
     debug("length:", film['length'], "sec")
-    nullin = open(os.devnull, 'r')
-    nullout = open(os.devnull, 'w')
     my_cmd = " ".join(TRANSITION_CMD + [film['name']])
+    debug("cmd:", my_cmd)
     proc = Popen(my_cmd, shell=True, preexec_fn=os.setsid, stdin=nullin, stdout=nullout)
     #sleep(film['length'])
     # save this process group id
     pgid = os.getpgid(proc.pid)
     debug("pgid:", pgid)
     pgid_last_transition = pgid
-    # close our null file handles
-    nullin.close()
-    nullout.close()
     debug("setting kill timer:", film['length'], "sec")
     threading.Timer(film['length'], stop_transition, [pgid]).start()
 
@@ -81,6 +83,29 @@ def stop_transition(pgid):
         pass
 
 
+def start_loop(film):
+    global pgid_last_loop
+    debug("filename:", film['name'])
+    debug("type", film['type'])
+    debug("start:", film['start'], "sec")
+    debug("end:", film['start']+film['length'], "sec")
+    debug("length:", film['length'], "sec")
+    # kill previous video before we start
+    if pgid_last_loop:
+        debug("previous loop still running:", pgid_last_loop, "(stopping in a sec)")
+        stop_content_in_a_sec(pgid_last_loop)
+    #if pgid_last_content:
+        #debug("previous content still running:", pgid_last_content, "(stopping in a sec)")
+        #stop_content_in_a_sec(pgid_last_content)
+    # start new video (saving the process handle)
+    my_cmd = " ".join(LOOP_CMD + [film['name']])
+    debug("cmd:", my_cmd)
+    proc = Popen(my_cmd, shell=True, preexec_fn=os.setsid, stdin=nullin, stdout=nullout)
+    # save this process group id
+    pgid = os.getpgid(proc.pid)
+    debug("pgid:", pgid)
+    pgid_last_loop = pgid
+
 def start_content(film):
     global pgid_last_content
     debug("filename:", film['name'])
@@ -88,32 +113,33 @@ def start_content(film):
     debug("start:", film['start'], "sec")
     debug("end:", film['start']+film['length'], "sec")
     debug("length:", film['length'], "sec")
-    nullin = open(os.devnull, 'r')
-    nullout = open(os.devnull, 'w')
     # kill previous video before we start
+    #if pgid_last_loop:
+        #debug("previous loop still running:", pgid_last_loop, "(stopping in a sec)")
+        #stop_content_in_a_sec(pgid_last_loop)
     if pgid_last_content:
-        debug("previous vid still running:", pgid_last_content, "(stopping in a sec)")
+        debug("previous content still running:", pgid_last_content, "(stopping in a sec)")
         stop_content_in_a_sec(pgid_last_content)
     # start new video (saving the process handle)
     my_cmd = " ".join(CONTENT_CMD + [film['name']])
-    proc = Popen(my_cmd, shell=True, preexec_fn=os.setsid, stdin=nullin, stdout=nullout)
+    debug("cmd:", my_cmd)
+    #proc = Popen(my_cmd, shell=True, preexec_fn=os.setsid, stdin=nullin, stdout=nullout)
+    proc = Popen(my_cmd, shell=True, preexec_fn=os.setsid, stdin=nullin)
     # save this process group id
     pgid = os.getpgid(proc.pid)
     debug("pgid:", pgid)
     pgid_last_content = pgid
-    #sleep(film['length'] - inter_video_delay)
-    # close our null file handles
-    nullin.close()
-    nullout.close()
-    debug("setting kill timer:", film['length'], "sec")
-    threading.Timer(film['length'], stop_content, [pgid]).start()
+    if (film['length']):
+        debug("setting kill timer:", film['length'], "sec")
+        threading.Timer(film['length'], stop_content, [pgid]).start()
 
 
 def stop_content(pgid):
     global pgid_last_content
     try:
         debug("Kiling", pgid)
-        os.killpg(pgid, signal.SIGTERM)
+        result = os.killpg(pgid, signal.SIGTERM)
+        debug("Result", result)
         pgid_last_content = None
     except:
         debug("Couldn't signal", pgid)
@@ -131,16 +157,36 @@ def main():
     #print content_film_list
 
     try:
+        loop = choice(loop_film_list)
+        start_loop(loop)
 
         while True:
+            max_content = len(content_film_list)-1
+            print ""
+            next_film = raw_input("Next video (0-%i): " % max_content)
+            try:
+                next_film = int(next_film)
+            except:
+                continue
+            if (next_film < 0 or next_film > max_content):
+                continue
             transition = choice(transition_film_list)
             start_transition(transition)
-            content = choice(content_film_list)
+            sleep(transition['length'] - inter_video_delay)
+            content = content_film_list[next_film]
             start_content(content)
             sleep(content['length'] - inter_video_delay)
+            transition = choice(transition_film_list)
+            start_transition(transition)
+            sleep(transition['length'] - inter_video_delay)
+            #start_loop(loop)
+            sleep(5)
 
     except KeyboardInterrupt:
         print "Done."
+        # close our null file handles
+        nullin.close()
+        nullout.close()
 
 if __name__ == "__main__":
     main()
